+++ src/model/Attendance.java
@@ -4,7 +4,7 @@ import java.sql.Date;
 
 public class Attendance {
     private int id;
-    private int userId;
+    private long userId;
     private String agenda;
     private Date attendanceDate;
     private String status;
@@ -12,7 +12,7 @@ public class Attendance {
     
     public Attendance() {}
     
-    public Attendance(int id, int userId, String agenda, Date attendanceDate, String status, String notes) {
+    public Attendance(int id, long userId, String agenda, Date attendanceDate, String status, String notes) {
         this.id = id;
         this.userId = userId;
         this.agenda = agenda;
@@ -24,8 +24,8 @@ public class Attendance {
     public int getId() { return id; }
     public void setId(int id) { this.id = id; }
     
-    public int getUserId() { return userId; }
-    public void setUserId(int userId) { this.userId = userId; }
+    public long getUserId() { return userId; }
+    public void setUserId(long userId) { this.userId = userId; }
     
     public String getAgenda() { return agenda; }
     public void setAgenda(String agenda) { this.agenda = agenda; }
