+++ src/model/Mentor.java
@@ -5,7 +5,7 @@ public class Mentor extends User {
         super();
     }
     
-    public Mentor(int id, String name, String email, String password) {
+    public Mentor(long id, String name, String email, String password) {
         super(id, name, email, password, "mentor");
     }
